Rework mode switching logic
The previous solution was not easily scalable and sucked in general.
Adding more modes would lead to a lot of code repetition and
progressively more work to implement every new mode. My new solution
with a global mode variable is much easier to manage.
Also it is much more reasonble.