FIX: make cell removal faster
Instead of removing the first child in a loop, now the resetGrid
function removes the last child until there's no more children.
The "Clear canvas" button works noticeably faster on grids with a
large amount of cells. The slider is also a bit less laggy when closer
to an upper limit.